HO: New improved handover algorithm
The handover algorithm selects the best cell, tries to solve congestion
of cells and forces handover in case of bad radio link:

Handover shall ensure that the mobile station is always on the best cell
during a call. This results in best speech quality and allows a lower
TX level (up-link) to save battery and reduce interference.

Sometimes this is not desirable, especially if a cell with large coverage
becomes congested after handover. In this situation it makes sense to
balance the usage between cells to prevent congestion. If a cell is
congested, one or several mobile stations need to be handovered to a
different cell, even if that cell is not the best cell for these mobile
station.

In case that a mobile station needs to handover to prevent a loss of
radio link, it makes sense to handover it to a cell that is congested
afterwards. In this situation, a different mobile station on the congested
target cell can be handovered to a non-congested neighbor cell, to solve
or balance the congestion.

Because congestion of TCH/F and TCH/H is handled individually or bad
quality might require change from TCH/H to TCH/F, assignment instead
of handover might be used to change the channel within the same BTS. This
new algorithm handles both assigment and handover depending on the target
BTS.

+++ openbsc/include/openbsc/gsm_data.h
@@ -96,7 +96,7 @@ struct neigh_meas_proc {
 	uint8_t bsic;
 	uint8_t rxlev[MAX_WIN_NEIGH_AVG];
 	unsigned int rxlev_cnt;
-	uint8_t last_seen_nr;
+	int last_seen_nr;
 };
 
 /* penalty timers for handover */
